backend-flask/requirements.txt
flask==3.0.0
openai==1.14.2
python-dotenv==1.0.0
